Move jdbc-to-googlecloud -> jdbc-and-googlecloud
* Eliminate JdbcIO related dependencies in googlecloud-to-googlecloud Templates
