+++ v2/jdbc-and-googlecloud/src/main/java/com/google/cloud/teleport/v2/templates/JdbcToPubSubAuto.java
