+++ v2/jdbc-and-googlecloud/src/main/resources/jdbc-to-bigquery-command-spec.json
