+++ v2/jdbc-and-googlecloud/src/main/resources/jdbc-to-pubsub-command-spec.json
