+++ v2/jdbc-and-googlecloud/src/main/resources/pubsub-to-jdbc-command-spec.json
