+++ v2/jdbc-and-googlecloud/src/test/java/com/google/cloud/teleport/v2/templates/JdbcToBigQueryReadWithPartitionsTest.java
