+++ v2/jdbc-and-googlecloud/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
