+++ v2/mysql-to-googlecloud/README_Jdbc_to_PubSub_Auto.md
@@ -79,7 +79,7 @@ mvn clean package -PtemplatesStage  \
 -DbucketName="$BUCKET_NAME" \
 -DstagePrefix="templates" \
 -DtemplateName="Jdbc_to_PubSub_Auto" \
--f v2/jdbc-to-googlecloud
+-f v2/jdbc-and-googlecloud
 ```
 
 
@@ -188,7 +188,7 @@ mvn clean package -PtemplatesRun \
 -DjobName="jdbc-to-pubsub-auto-job" \
 -DtemplateName="Jdbc_to_PubSub_Auto" \
 -Dparameters="driverClassName=$DRIVER_CLASS_NAME,connectionUrl=$CONNECTION_URL,username=$USERNAME,password=$PASSWORD,driverJars=$DRIVER_JARS,connectionProperties=$CONNECTION_PROPERTIES,query=$QUERY,KMSEncryptionKey=$KMSENCRYPTION_KEY,partitionColumn=$PARTITION_COLUMN,table=$TABLE,numPartitions=$NUM_PARTITIONS,lowerBound=$LOWER_BOUND,upperBound=$UPPER_BOUND,outputTopic=$OUTPUT_TOPIC" \
--f v2/jdbc-to-googlecloud
+-f v2/jdbc-and-googlecloud
 ```
 
 ## Terraform
@@ -205,7 +205,7 @@ To use the autogenerated module, execute the standard
 [terraform workflow](https://developer.hashicorp.com/terraform/intro/core-workflow):
 
 ```shell
-cd v2/jdbc-to-googlecloud/terraform/Jdbc_to_PubSub_Auto
+cd v2/jdbc-and-googlecloud/terraform/Jdbc_to_PubSub_Auto
 terraform init
 terraform apply
 ```
