+++ v2/postgresql-to-googlecloud/README_Jdbc_to_PubSub.md
@@ -79,7 +79,7 @@ mvn clean package -PtemplatesStage  \
 -DbucketName="$BUCKET_NAME" \
 -DstagePrefix="templates" \
 -DtemplateName="Jdbc_to_PubSub" \
--f v2/jdbc-to-googlecloud
+-f v2/jdbc-and-googlecloud
 ```
 
 
@@ -179,7 +179,7 @@ mvn clean package -PtemplatesRun \
 -DjobName="jdbc-to-pubsub-job" \
 -DtemplateName="Jdbc_to_PubSub" \
 -Dparameters="driverClassName=$DRIVER_CLASS_NAME,connectionUrl=$CONNECTION_URL,username=$USERNAME,password=$PASSWORD,driverJars=$DRIVER_JARS,connectionProperties=$CONNECTION_PROPERTIES,query=$QUERY,outputTopic=$OUTPUT_TOPIC,KMSEncryptionKey=$KMSENCRYPTION_KEY,disabledAlgorithms=$DISABLED_ALGORITHMS,extraFilesToStage=$EXTRA_FILES_TO_STAGE" \
--f v2/jdbc-to-googlecloud
+-f v2/jdbc-and-googlecloud
 ```
 
 ## Terraform
@@ -196,7 +196,7 @@ To use the autogenerated module, execute the standard
 [terraform workflow](https://developer.hashicorp.com/terraform/intro/core-workflow):
 
 ```shell
-cd v2/jdbc-to-googlecloud/terraform/Jdbc_to_PubSub
+cd v2/jdbc-and-googlecloud/terraform/Jdbc_to_PubSub
 terraform init
 terraform apply
 ```
